get post by id endpoint

+++ src/main/java/com/blog/app/blog/exception/ResourceNotFoundException.java
@@ -12,9 +12,9 @@ public class ResourceNotFoundException extends RuntimeException{
 
     private String resourceName;
     private String fieldName;
-    private int fieldValue;
+    private long fieldValue;
 
-    public ResourceNotFoundException(String resourceName, String fieldName, int fieldValue) {
+    public ResourceNotFoundException(String resourceName, String fieldName, long fieldValue) {
         super(String.format("%s not found with %s : %s",resourceName,fieldName,fieldValue)); // post not found with id : 12
         this.resourceName = resourceName;
         this.fieldName = fieldName;
